replace DirectoryNotFoundException with logic check

+++ StrategyPattern/Commands/CreateTxtCommand.cs
@@ -11,18 +11,13 @@ namespace StrategyPattern
     {
         public void Process(string param)
         {
-            try
+            if (!Directory.Exists(param))
             {
-                File.Create(param + "\\test.txt");
+                Console.WriteLine("directory {0} doesn't exist", param);
+                return;
             }
-            catch (DirectoryNotFoundException ex)
-            {
-                Console.WriteLine(ex.Message);
-            }
-            catch (PathTooLongException ex)
-            {
-                Console.WriteLine(ex.Message);
-            }
+
+            File.Create(param + "\\test.txt");
         }
     }
 }
